backend/requirements.txt
alembic==1.16.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
bcrypt==3.2.0
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
ecdsa==0.19.0
fastapi==0.115.7
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.10
MarkupSafe==3.0.2
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.20
redis==5.2.1
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.2
typing_extensions==4.12.2
uvicorn==0.34.0
Werkzeug==3.1.3
black 
ruff 
flake8